feat(dispatcher): stream_frame codec + Dispatcher multiplexer; OTA rides them
Extracts the frame codec that lived in ota/detail/ota_stream_protocol.hpp into a
new dependency-free 'stream_frame' component (magic/type/len/crc framing, CRC-32,
put/get helpers, incremental resynchronizing StreamParser; Frame::type is now a
generic uint8_t). ota_stream_protocol.hpp becomes a thin facade that re-exports
those symbols under the historical espp::detail::ota_stream namespace and keeps
the OTA MessageType enum + make_*/parse_* helpers, so existing OTA (and the open
coredump branch) keep compiling unchanged.

Adds a new 'dispatcher' component: espp::Dispatcher parses one stream once and
routes each frame to a per-module handler by the type byte's high-nibble module
id (module_of = type >> 4). This lets several framed protocols share one USB /
socket / UART link instead of running a StreamParser per protocol. The module id
math is backward compatible with the deployed wire codes: OTA opcodes 0x0X/0x8X
-> module 0, coredump 0x4X/0xCX -> module 4 (the 0x80 reply bit sits in the high
nibble, so requests and replies map to disjoint module ids and a device-side
dispatcher only ever sees the request modules).

The ota example now feeds a Dispatcher with OTA registered on module 0 (instead
of a bare StreamParser), demonstrating the pattern and cleanly ignoring other
protocols' frames rather than replying 'unknown message type'.

Host tests: components/ota/test (codec, updated for uint8_t Frame::type) and a
new components/dispatcher/test (routing / coexistence / reset) both pass. ota
example builds clean (esp32s3). Docs + Doxygen inputs added for both components.

The ota + dispatcher manifests point espp/stream_frame at the local source via
override_path until it is published (mirroring lilygo-t5-47's bq27220/pca9535).

+++ components/ota/test/ota_protocol_host_test.cpp
@@ -1,6 +1,6 @@
 // Host-buildable unit tests for the espp OTA stream protocol framing. Build &
 // run with:
-//   c++ -std=c++20 -Werror -I components/ota/include \
+//   c++ -std=c++20 -Werror -I components/ota/include -I components/stream_frame/include \
 //       components/ota/test/ota_protocol_host_test.cpp -o test && ./test
 //
 // These tests exercise espp::detail::ota_stream directly so they need no
@@ -90,23 +90,23 @@ static void test_round_trip_all_types() {
   if (frames.size() != 7)
     return;
 
-  CHECK(frames[0].type == MessageType::Begin);
+  CHECK(frames[0].type == static_cast<uint8_t>(MessageType::Begin));
   CHECK(ota::parse_u32_payload(frames[0]).value_or(0) == 1234567u);
 
-  CHECK(frames[1].type == MessageType::Data);
+  CHECK(frames[1].type == static_cast<uint8_t>(MessageType::Data));
   CHECK(frames[1].payload.size() == sizeof(image_bytes));
   CHECK(std::memcmp(frames[1].payload.data(), image_bytes, sizeof(image_bytes)) == 0);
 
-  CHECK(frames[2].type == MessageType::Finish);
+  CHECK(frames[2].type == static_cast<uint8_t>(MessageType::Finish));
   CHECK(frames[2].payload.empty());
 
-  CHECK(frames[3].type == MessageType::Abort);
+  CHECK(frames[3].type == static_cast<uint8_t>(MessageType::Abort));
   CHECK(frames[3].payload.empty());
 
-  CHECK(frames[4].type == MessageType::Ok);
+  CHECK(frames[4].type == static_cast<uint8_t>(MessageType::Ok));
   CHECK(ota::parse_u32_payload(frames[4]).value_or(0) == 6u);
 
-  CHECK(frames[5].type == MessageType::Error);
+  CHECK(frames[5].type == static_cast<uint8_t>(MessageType::Error));
   const auto err = ota::parse_error(frames[5]);
   CHECK(err.has_value());
   if (err.has_value()) {
@@ -114,7 +114,7 @@ static void test_round_trip_all_types() {
     CHECK(err->message == "flash write failed");
   }
 
-  CHECK(frames[6].type == MessageType::Progress);
+  CHECK(frames[6].type == static_cast<uint8_t>(MessageType::Progress));
   const auto prog = ota::parse_progress(frames[6]);
   CHECK(prog.has_value());
   if (prog.has_value()) {
@@ -140,9 +140,9 @@ static void test_split_across_chunks() {
   CHECK(frames.size() == 2);
   CHECK(parser.dropped_bytes() == 0);
   if (frames.size() == 2) {
-    CHECK(frames[0].type == MessageType::Data);
+    CHECK(frames[0].type == static_cast<uint8_t>(MessageType::Data));
     CHECK(frames[0].payload.size() == sizeof(data_bytes));
-    CHECK(frames[1].type == MessageType::Ok);
+    CHECK(frames[1].type == static_cast<uint8_t>(MessageType::Ok));
     CHECK(ota::parse_u32_payload(frames[1]).value_or(0) == 10u);
   }
 
@@ -170,7 +170,7 @@ static void test_resync_on_corruption() {
   const auto frames = parser.feed(stream);
   CHECK(frames.size() == 1);
   if (frames.size() == 1) {
-    CHECK(frames[0].type == MessageType::Ok);
+    CHECK(frames[0].type == static_cast<uint8_t>(MessageType::Ok));
     CHECK(ota::parse_u32_payload(frames[0]).value_or(0) == 4u);
   }
   CHECK(parser.dropped_bytes() > 0);
@@ -195,7 +195,7 @@ static void test_oversized_len_rejected() {
   const auto frames = parser.feed(stream);
   CHECK(frames.size() == 1);
   if (frames.size() == 1)
-    CHECK(frames[0].type == MessageType::Finish);
+    CHECK(frames[0].type == static_cast<uint8_t>(MessageType::Finish));
   CHECK(parser.dropped_bytes() > 0);
   // The oversized length must never be buffered/waited for.
   CHECK(parser.buffered() < ota::kMaxFrameSize);
@@ -217,14 +217,14 @@ static void test_oversized_len_rejected() {
 static void test_malformed_reply_payloads() {
   std::printf("test_malformed_reply_payloads\n");
   // Wrong-size payloads must be rejected by the typed parse helpers.
-  Frame f{MessageType::Ok, {0x01, 0x02}};
+  Frame f{static_cast<uint8_t>(MessageType::Ok), {0x01, 0x02}};
   CHECK(!ota::parse_u32_payload(f).has_value());
-  Frame e{MessageType::Error, {0x01, 0x02, 0x03}};
+  Frame e{static_cast<uint8_t>(MessageType::Error), {0x01, 0x02, 0x03}};
   CHECK(!ota::parse_error(e).has_value());
-  Frame p{MessageType::Progress, {0x01, 0x02, 0x03, 0x04}};
+  Frame p{static_cast<uint8_t>(MessageType::Progress), {0x01, 0x02, 0x03, 0x04}};
   CHECK(!ota::parse_progress(p).has_value());
   // An ERROR with just a code (no message) is valid.
-  Frame e2{MessageType::Error, {0x05, 0x00, 0x00, 0x00}};
+  Frame e2{static_cast<uint8_t>(MessageType::Error), {0x05, 0x00, 0x00, 0x00}};
   const auto info = ota::parse_error(e2);
   CHECK(info.has_value());
   if (info.has_value()) {
